+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning].
 - fixed recipe viewer integration endpoints for Fabric
 - fixed unnecessary memory usage for debug handler
 - fixed Mekanism recipe unifier using wrong recipe keys
+- fixed EnderIO Sag Mill recipe output unification causing serialization failures
 
 ## [1.1.0] - 2024-09-27